Issue resolved: Multiple StatWriter
The StatWriters were not removed after running one iteration of optimisation run.
It caused multiple StatWriters and the all trial tensorboard results being overwritten with the latest trial run.

Another related issue was the non-closure of the running SummaryWriters.
Multiple threads were simultaneously running, ending with a pickling error as it exceeds the number of threads on
machine.